[EI-254] [web] feat: Move chat component to domain folder
